+++ dec_9/src/main.rs
@@ -1,14 +1,14 @@
 fn main() {
-	let input: Vec<usize> = include_str!("input.txt")
+	let input: Vec<u64> = include_str!("input.txt")
 		.lines()
-		.map(|x| x.parse::<usize>().unwrap())
+		.map(|x| x.parse::<u64>().unwrap())
 		.collect::<Vec<_>>();
 	let sol1 = solve1(&input);
 	let sol2 = solve2(sol1, &input);
 	println!("{} {}", sol1, sol2);
 }
 
-fn solve1(nums: &[usize]) -> usize {
+fn solve1(nums: &[u64]) -> u64 {
 	for i in 0..(nums.len() - 26) {
 		let nums = &nums[i..];
 		if !is_sum_of(nums[25], &nums[..25]) {
@@ -18,7 +18,7 @@ fn solve1(nums: &[usize]) -> usize {
 	unreachable!();
 }
 
-fn solve2(target: usize, nums: &[usize]) -> usize {
+fn solve2(target: u64, nums: &[u64]) -> u64 {
 	for i in 0..nums.len() {
 		let mut sum = 0;
 		for j in 0.. {
@@ -27,7 +27,7 @@ fn solve2(target: usize, nums: &[usize]) -> usize {
 				let range = &nums[i..=i + j];
 				let max = *range.iter().max().unwrap_or(&0);
 				let min = *range.iter().min().unwrap_or(&0);
-				assert!(range.iter().sum::<usize>() == target);
+				assert!(range.iter().sum::<u64>() == target);
 				return max + min;
 			}
 			if sum > target {
@@ -38,7 +38,7 @@ fn solve2(target: usize, nums: &[usize]) -> usize {
 	0
 }
 
-fn is_sum_of(n: usize, nums: &[usize]) -> bool {
+fn is_sum_of(n: u64, nums: &[u64]) -> bool {
 	nums.iter()
 		.enumerate()
 		.any(|(i, &lhs)| nums.iter().skip(i + 1).any(|&rhs| lhs + rhs == n))
